requirements.txt
gradio==5.36.2
pandas>=2.3.1
numpy>=2.1.2
PyMuPDF>=1.26.3
openai>=1.93.3
python-dotenv>=1.1.1
# Added dependencies
scikit-learn>=1.7.0
langchain-text-splitters>=0.3.8
tiktoken>=0.9.0
